Update Multiple Regression
In this project, the investors can have an insight into the results. They will found out in which startup invest their money.
In the linear regression model, the following results can be found :
Which independent variable has the highest effect on the profit
Predict future profits for new startups based on the same information.